requirements.txt
pydantic
pydantic-settings
typing
openai
dotenv-python